app/requirements.txt
flask==2.1.2
gunicorn==20.1.0
flask-wtf==1.0.1
google-cloud-texttospeech==2.11.1
google-cloud-storage==2.4.0
google-cloud-firestore==2.5.3
flask-bootstrap==3.3.7.1
flask-cors==3.0.10

